+++ test/computer_test.dart
@@ -4,26 +4,26 @@ import 'package:test/test.dart';
 
 void main() {
   test('Computer turn on', () async {
-    final computer = Computer();
+    final computer = Computer.create();
     await computer.turnOn();
     expect(computer.isRunning, equals(true));
     await computer.turnOff();
   });
 
   test('Computer initially turned off', () async {
-    final computer = Computer();
+    final computer = Computer.create();
     expect(computer.isRunning, equals(false));
   });
 
   test('Computer turn off', () async {
-    final computer = Computer();
+    final computer = Computer.create();
     await computer.turnOn();
     await computer.turnOff();
     expect(computer.isRunning, equals(false));
   });
 
   test('Computer reload', () async {
-    final computer = Computer();
+    final computer = Computer.create();
     await computer.turnOn();
     expect(computer.isRunning, equals(true));
     await computer.turnOff();
@@ -35,7 +35,7 @@ void main() {
   });
 
   test('Execute function with param', () async {
-    final computer = Computer();
+    final computer = Computer.create();
     await computer.turnOn();
 
     expect(await computer.compute<int, int>(fib, param: 20), equals(fib(20)));
@@ -44,7 +44,7 @@ void main() {
   });
 
   test('Stress test', () async {
-    final computer = Computer();
+    final computer = Computer.create();
     await computer.turnOn();
 
     const numOfTasks = 500;
@@ -67,7 +67,7 @@ void main() {
   });
 
   test('Execute function without params', () async {
-    final computer = Computer();
+    final computer = Computer.create();
     await computer.turnOn();
 
     expect(await computer.compute<int, int>(fib20), equals(fib20()));
@@ -76,7 +76,7 @@ void main() {
   });
 
   test('Execute static method', () async {
-    final computer = Computer();
+    final computer = Computer.create();
     await computer.turnOn();
 
     expect(
@@ -88,7 +88,7 @@ void main() {
   });
 
   test('Execute async method', () async {
-    final computer = Computer();
+    final computer = Computer.create();
     await computer.turnOn();
 
     expect(
@@ -100,7 +100,7 @@ void main() {
   });
 
   test('Error method', () async {
-    final computer = Computer();
+    final computer = Computer.create();
     await computer.turnOn();
 
     expect(
@@ -111,11 +111,18 @@ void main() {
     await computer.turnOff();
   });
 
-  test('Computer is a singleton', () async {
-    final computer1 = Computer();
-    final computer2 = Computer();
+  test('Computer instance is a singleton', () async {
+    final computer1 = Computer.shared();
+    final computer2 = Computer.shared();
 
-    expect(computer1 == computer2, equals(true));
+    expect(identical(computer1, computer2), true);
+  });
+
+  test('Computer create create new instances', () async {
+    final computer1 = Computer.create();
+    final computer2 = Computer.create();
+
+    expect(identical(computer1, computer2), false);
   });
 }